110: Controllers: TestUtils moved to new class
