+++ src/test/java/pl/kacperk/pokemonservicefullstack/controller/AuthControllerTest.java
@@ -21,12 +21,12 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.MATCHING_PASS_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.PASS_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTERED_USER_NAME;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTERED_USER_PASS;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTER_REQUEST_DTO;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.getLoggedUserSession;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.MATCHING_PASS_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.PASS_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTERED_USER_NAME;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTERED_USER_PASS;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTER_REQUEST_DTO;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.getLoggedUserSession;
 import static pl.kacperk.pokemonservicefullstack.controller.AuthController.ERROR_MSG_ATTR;
 import static pl.kacperk.pokemonservicefullstack.controller.AuthController.LOGIN_ERROR_MAPPING;
 import static pl.kacperk.pokemonservicefullstack.controller.AuthController.LOGIN_MAPPING;
