+++ src/test/java/pl/kacperk/pokemonservicefullstack/controller/PokemonControllerTest.java
@@ -30,15 +30,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.EVOLUTIONS_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.ID_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.LOGIN_URL;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.NAME_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.PHOTO_URL_PROP;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTERED_USER_NAME;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTERED_USER_PASS;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.REGISTER_REQUEST_DTO;
-import static pl.kacperk.pokemonservicefullstack.TestUtils.ControllerUtils.getLoggedUserSession;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.EVOLUTIONS_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.ID_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.LOGIN_URL;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.NAME_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.PHOTO_URL_PROP;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTERED_USER_NAME;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTERED_USER_PASS;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.REGISTER_REQUEST_DTO;
+import static pl.kacperk.pokemonservicefullstack.controller.ControllerTestUtils.getLoggedUserSession;
 import static pl.kacperk.pokemonservicefullstack.TestUtils.PageableUtils.DEF_FIELD_TO_SORT;
 import static pl.kacperk.pokemonservicefullstack.TestUtils.PageableUtils.DEF_MATCH;
 import static pl.kacperk.pokemonservicefullstack.TestUtils.PageableUtils.DEF_PAGE_NUM;
